requirements.txt
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
openenv-core>=0.2,<1.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
openai>=2.7.2,<3.0
requests>=2.32,<3.0
PyYAML>=6.0,<7.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
httpx>=0.27,<1.0
anyio>=4.0,<5.0
